requirements.txt
BAC0>=0.99
python-dotenv>=1.0,<2
Flask>=2.3,<4
pywebview>=4.4
psutil>=5.9
